set up reducer and Redux store
