--- Anywhere-Fitness/src/reducers/reducer.js
@@ -1,17 +1,0 @@
-const initialState = {
-    isLoading: false,
-    registration: {
-        "username": "",
-        "email": "",
-        "password": "",
-        "role": ""
-    },
-    error: false
-}
-
-export const reducer = (state = initialState, action) => {
-    switch(action.type) {
-        default:
-            return state;
-    }
-}
